pyproject.toml
[project]
name = "whisper-api"
version = "0.1.0"
description = "Add your description here"
authors = [
    { name = "Sogo Kato", email = "sogo.kato@outlook.com" }
]
dependencies = [
    "openai-whisper @ git+https://github.com/openai/whisper.git",
    "fastapi>=0.115.2",
    "uvicorn>=0.31.1",
    "python-multipart>=0.0.12",
    "ffmpeg-python>=0.2.0",
    "pydantic-settings>=2.6.0",
]
readme = "README.md"
requires-python = ">= 3.8"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = [
    "pyaudio>=0.2.14",
]

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/whisper_api"]
